[triton_op] fix autotuning
The problem was we were shoving SymInts into the constant_args side
table. The root problem is that torch.fx.node.base_types, which we use
to determine what can be put in the graph, doesn't actually have SymInt
in it. This PR fixes base_types to include SymInt.

Test Plan:
- tests

ghstack-source-id: b5fbeafb811f7baec74ee6544fff6f7fbb276626
Pull Request resolved: https://github.com/pytorch/pytorch/pull/131363